+++ f17.c
@@ -1,31 +1,11 @@
 #include "math.h"
 #include "f17.h"
 #include "matrix.h"
-static double L1_norm(int n, double *a)
-{
-    int i,j;
-    double sum, max=-1;
-    
-    for(j=0;j<n;j++)
-    {
-        sum=0;
-        for(i=0;i<n;i++) sum+=fabs(a[i*n+j]);
-        if (sum>max) max=sum;
-    }
-    
-    return max;
-}
 
-int solver(int n, double *a, double *b, double *x, int *ord)
+int solver(int n, double *a, double *b, double *x, int *ord, double norm)
 {
     int i,j,k,max_i, max_j;
-    double max_a, buf, norm;
-    
-    norm=L1_norm(n,a);
-    // printf("%lf\n",norm);
-    if(fabs(norm)<1e-100) return -1;
-
-    for(i=0;i<n;i++) ord[i]=i;
+    double max_a, buf;
     
     for(k=0;k<n;k++)
     {
@@ -65,32 +45,25 @@ int solver(int n, double *a, double *b, double *x, int *ord)
 	}
         
         
-        //  printf("%e\n\n",a[k*n+k]);
-        //  printf("%e\n\n",norm*1e-16);
                 
         if (fabs(a[k*n+k])<(1.269653e-16*norm)) {return -1;}
         
         buf=1.0/a[k*n+k];
         
-       // a[k*n+k]=1.0;
         
         for(j=k+1;j<n;j++) a[k*n+j]*=buf;
         b[k]*=buf;
 
         for(i=0;i<k;i++)
         {
-            //if (k==i) continue;
             buf=a[i*n+k];
-            //a[i*n+k]=0.0;
             for (j=k+1;j<n;j++) a[i*n+j]-=a[k*n+j]*buf;
             b[i]-= b[k]*buf;
         }
         
         for(i=k+1;i<n;i++)
         {
-            //if (k==i) continue;
             buf=a[i*n+k];
-            //a[i*n+k]=0.0;
             for (j=k+1;j<n;j++) a[i*n+j]-=a[k*n+j]*buf;
             b[i]-= b[k]*buf;            
         }
