+++ f17.h
@@ -1,1 +1,1 @@
-int solver(int n, double *a, double *b, double *x, int *ord);
+int solver(int n, double *a, double *b, double *x, int *ord, double norm);
